requirements.txt
boto3==1.22.10
botocore==1.25.10
Brotli==1.0.9
CacheControl==0.12.11
certifi==2021.10.8
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
dash==1.13.4
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
Flask==1.1.2
Flask-Assets==2.0
Flask-Compress==1.5.0
future==0.18.2
idna==2.10
iso3166==2.0.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==1.0.0
mapbox==0.18.0
MarkupSafe==1.1.1
msgpack==1.0.3
numpy==1.19.1
pandas==1.0.5
plotly==4.9.0
polyline==1.4.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
retrying==1.3.3
s3transfer==0.5.2
six==1.16.0
uritemplate==4.1.1
urllib3==1.25.11
webassets==2.0
Werkzeug==1.0.1
